backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
pyspark>=3.4.0
python-multipart>=0.0.1